move APIFlowResposne model into client module
