+++ tests/tools/flow/test_model.py
@@ -1,7 +1,6 @@
 from typing import Any
 
-from keboola_mcp_server.client import ORCHESTRATOR_COMPONENT_ID
-from keboola_mcp_server.tools.flow.api_models import APIFlowResponse
+from keboola_mcp_server.client import ORCHESTRATOR_COMPONENT_ID, APIFlowResponse
 from keboola_mcp_server.tools.flow.model import (
     Flow,
     FlowConfiguration,
